Added respect in the MARS downloader for delete tempfiles property

+++ icenet2/data/interfaces/mars.py
@@ -154,9 +154,11 @@ retrieve,
                 if daily_path not in self._files_downloaded:
                     self._files_downloaded.append(daily_path)
 
-        logging.info("Removing {}".format(request_target))
         ds.close()
-        os.unlink(request_target)
+
+        if self.delete:
+            logging.info("Removing {}".format(request_target))
+            os.unlink(request_target)
 
     def download(self):
         logging.info("Building request(s), downloading and daily averaging "
